Updated README to include new scripts
I tested the new scripts that could be run on windows, thus prd-pack-mac and prd-pack-linux need to be tested.

Test command doesn't work due to no test existing, it can be removed but I have left it in for now.

Don't know how to describe lint command, please append to the description if you know what it does.

All other commands have been tested and work on windows platform.